requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.3
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.1
contourpy==1.2.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.13.1
fonttools==4.47.2
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm_watsonx_ai==1.1.14
idna==3.6
imageio==2.34.0
importlib_metadata==8.5.0
ipykernel==6.29.0
ipython==8.21.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.7.0
jmespath==1.0.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lightning-utilities==0.10.1
lomond==0.3.3
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mplcyberpunk==0.7.1
mpmath==1.3.0
mysql-connector==2.2.9
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
openai==1.54.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
parso==0.8.3
pdfkit==1.0.0
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.17.2
pyparsing==3.1.1
pyserial==3.5
PySocks==1.7.1
pyTelegramBotAPI==4.23.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-telegram-bot==21.7
pytz==2024.1
pyzmq==25.1.2
requests==2.32.2
scapy==2.6.1
scikit-learn==1.4.0
scipy==1.12.0
selenium==4.21.0
setuptools==58.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
torch==2.2.0
torchaudio==2.2.0
torchmetrics==1.3.1
torchsummary==1.5.1
torchvision==0.17.0
tornado==6.4
tqdm==4.67.0
traitlets==5.14.1
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2023.4
urllib3==2.2.0
wcwidth==0.2.13
wget==3.2
wheel==0.45.1
wkhtmltopdf==0.2
wsproto==1.2.0
zipp==3.20.2
